repair: moved component search-field
